feat(dasclaw_core): extract JobError to dasclaw_core::error (ADR-152 §3 F3.6 slice 2)
Verbatim port per ADR-129 §1.3. Sets up dasclaw_core::error module so the
upcoming context::manager slice can depend on JobError from a stable
location.

Changes:
- crates/dasclaw_core/src/error.rs (new): JobError enum + job_error_display unit test, byte-for-byte copy of the original.
- crates/dasclaw_core/src/lib.rs: pub mod error; pub use error::JobError;
- desktop-client/ironclaw/src/error.rs: replaced JobError definition with pub use dasclaw_core::JobError shim (mirrors WorkspaceError / ToolError pattern). Removed the duplicated job_error_display test (kept in dasclaw_core). Dropped now-unused std::time::Duration import.

Top-level Error::Job(#[from] JobError) is unchanged — the re-exported type is identical, so #[from] continues to work (covered by top_level_error_from_conversions).

Closes #717
Refs #632 (ADR-152 §3 F3.6 umbrella — stays open until all 6 slices land)
